Correction suppression Tier

+++ app/Livewire/Tiers/DetailTiers.php
@@ -17,6 +17,7 @@ use App\Models\factureClientLigne;
 use App\Models\CommandeClientEntete;
 use App\Models\CommandeClientLigne;
 use App\Models\SoldeTier;
+use App\Models\Opportunite;
 
 class DetailTiers extends Component
 {
@@ -296,20 +297,33 @@ class DetailTiers extends Component
             $autoriser = $role[0]->supprimer_tier;
             if($autoriser == 1){   
                 if($id){
-                    Tier::where('id',$id)->delete();
-                    SoldeTier::where('id_tier',$id)->delete();                                         
-                    $id_activite = $id;
-                    $page = 'Tiers';
-                    LogActivity::addToLog('Tier supprimé', $id_activite, $page);
-                    $this->dispatch('alert',                    
-                        title:'Suppression effectuée!',
-                        timer:3000,
-                        icon:'success',
-                        toast:true,
-                        showConfirmButton: false,
-                        position:'top-end',
-                    ); 
-                    $this->redirect('/listing-tiers?active=3&champ=3-2', navigate: true); 
+                    $test_opport = Opportunite::where('societe',auth()->user()->societe)->where('id_client',$id)->count();
+                    if($test_opport == 0){
+                        Tier::where('id',$id)->delete();
+                        SoldeTier::where('id_tier',$id)->delete();                                         
+                        $id_activite = $id;
+                        $page = 'Tiers';
+                        LogActivity::addToLog('Tier supprimé', $id_activite, $page);
+                        $this->dispatch('alert',                    
+                            title:'Suppression effectuée!',
+                            timer:3000,
+                            icon:'success',
+                            toast:true,
+                            showConfirmButton: false,
+                            position:'top-end',
+                        ); 
+                        $this->redirect('/listing-tiers?active=3&champ=3-2', navigate: true); 
+                    }
+                    else{                        
+                        $this->dispatch('alert',                    
+                            title:'Désolé, vous ne pouvez pas supprimer un tier lié à une opportunité!',
+                            timer:5000,
+                            icon:'error',
+                            toast:true,
+                            showConfirmButton: false,
+                            position:'top-end',
+                        );       
+                    } 
                 }
             } 
             else{
